Move some data files that are currently used by the package into a data/ subdirectory of the package for easier finding later.
In the long term these files will probably go away but we'll see.

Remove the sqlalchemy_example_auth.db from the repository; there's no
good reason to have it here (and in fact it's not even up-to-date with
the current ORM model, but nevermind).  Remove the test/query_authdb.py
module which uses it (as a "test" it's trivial and not useful).