Clean up of the imports in each module, including correct use of relative imports.
Also made newssourceaggregator.prototypes a proper package so that it
can function correctly, though it will probably be removed soon.

This includes a workaround for a bug in the databasetools.mongo module,
and a new utility function `traverse_dict` to replace getDictSubElement
which was supposed to be in the systemtools package but now isn't.