[backend] Initial prototype controller and RSS feed crawler.
It doesn't do anything but periodically queue RSS feeds (from a pre-provided
list in the database, later we will add means of expanding that list both
manually, and from scrape results), and then download and parse those feeds to
obtain URLs for articles.  Articles are not yet crawled/scraped.  That will be
next.

To start up RabbitMQ and MongoDB servers for development using Docker, you
can use the provided docker-compose.yml like:

    $ docker-compose up

The controller can be started by running:

    $ python -m renewal_backend.controller

For now there is assumed to be only one controller (though in principle this
could also be sharded).  Any number of feed crawler agents can be started by
running:

    $ python -m renewal_controller.crawlers.feed

Feed URLs will be distributed across crawler agents.

There are several TODO notes in the code indicating further refactoring that
is needed, as the architecture has become clearer in my mind.  I would also
like to add separate queues/routing keys for sending new article URLs to the
controller (for saving in the database) as opposed to sending new article URLs
directly on to crawlers to begin working on them immediately before having to
wait for the controller.  It's not clear whether or not this is necessary but
I suspect it will keep data moving through the system at a smoother pace.