[backend] Rename the 'resources' collection and exchange to 'articles' respectively.
Under my current data model an article is like a sub-type of the more
generic 'resource' type.  However, this makes it more explicit that the
product of crawling feeds is not a generic resource, but specifically new
articles.

Originally we did consider having a generic 'urls' collection and a many to one
mapping of urls to articles, since the same URL in some cases may reference the
same article.  For example, once a duplicate article is detected, we can
(arbitrarily?) pick one copy to server as the canonical instance, and for all
duplicates add some 'duplicate_of' property that references the canonical
instance.