[backend] Start adding schemas for some of the known data models in the system.
Support schema validation in MongoDB.  MongoDB uses its own variant of JSON
Schema that cannot be said to be a superset, since although it adds its own
extensions, it also does *not* support some standard JSON schema features.
For now, since this is only used to validate documents inserted into MongoDB
(as well as to document the data model) we stick with the MongoDB variant only.