[backend] Some refactoring of how exchanges used by each Agent are initialized.
Still not totally happy with this, but it does reduce some churn: Previously we
had to call declare_exchange for each Producer instance created, even if we
have multiple Producers using the same exchange.  Now, all exchanges used
by the Agent are declared in one shot over a dedicated channel.