[app][refactoring] properly implement async account sign-in using redux-thunk
as noted in one of my comments, the boilerplate needed to implement async
actions using thunks can be quite verbose, and I have some ideas for how
to reduce it (I believe there are some existing packages for this as well, but
I think I can write my own more easily--though I will check if the redux-thunk
package itself includes any helpers).  Nevertheless, doing it the "verbose"
way is helpful to understand how it works.

now, in order to initiate an account sign-in, the app doesn't call
signInAnonymous directly--instead it calls a signIn() action creator which
creates the thunk action--this in turn takes care of calling signInAnonymous()
and handling its response when it resolves (no more using the onAuthStateChanged
callback; rather we handle the response from signInAnonymous() directly)

also improved the splash screen--it displays the RENEWAL logo nicely, and
leaves it up until authentication completes succesfully; will have to add
error handling here as well, but later