[app][refactoring] added some stubs for handling settings in the redux state and rewrote the settings page
I kept the settings page looking almost exactly as it did before, but reduced
a lot of duplication and messiness; parts of that didn't work before still
don't work properly (such as requesting user permissions for sensors).

I don't want to spend a lot of time making the rest work right now, since
this is still kind of schematic--none of these settings are actually used
in the app yet (nor were they before I started playing with it), but the
foundations for the settings page are much better now.  It's easy to add
and remove settings declaratively.