[app][enhancement] get animated header working with WebView--this was much trickier than expected due to the way WebViews are sized within ScrollViews
(I actually thought WebView was based on ScrollView but it isn't, at least
not directly).  Had to pull out some fancy tricks.

The AutoHeightWebView works but is a little bit janky because sometimes the
height doesn't change--and hence the component re-rendered--quickly enough to
display the additional contents.  But as far as I can tell this is only really
a problem when scrolling down the page very fast--not reading, or even skimming,
at a normal pace.