[app][refactoring] reworked app configuration to use the system provided by Expo for this, rather than dotenv files and the custom config.js
see https://docs.expo.io/workflow/configuration/

the added app.config.js looks for a special app.config.dev.json when in
development mode--things will sort of work if this doesn't exist, but it's
highly encouraged to create it first (I will add docs on how to do this).

for other build environments (staging, production) it is *required* to have
an associated app.config.<environment>.json, such as app.config.prod.json for
production.  this environment name can be anything, but it's determined by
the string passed as the --release-channel argument to `expo publish` or
`expo build:android`, `expo build:ios`, etc.