[backend][bug] fix a pretty severe misunderstanding I had with RabbitMQ
I actually knew this was wrong but for some reason I wasn't thinking about it.

All updates to resources (feeds, articles, images) were being sent to the same
queue named update_resource.  update_resource was bound to all three exchanges
of the same name: we have a 'feeds' exchange, and 'articles' exchange and an
'images' exchange, but just one 'update_resource' queue bound to all 3
exchanges.

At some point it entered my thinking that there would somehow be a separate
"update_resource" queue, one bound to each exchange, but that's not how it
works at all.  Queues are independent of exchanges and the same queue can
be bound to one or many exchages.

As a result of this, update_resource calls would be handled by an arbitrary
update_resource worker at random, but each worker is bound to a specific
resource type.  This resulted in failed resource updates if an update to the
wrong resource type is sent to the wrong worker.

This fixes it by making sure there is a separate queue for updating each
type of resource (feeds, articles, images).