[backend][task] first iteration of a site scraper service using newspaper3k
the hook system for update_resource is a bit of a hack; it might almost be
easier to just implement a separate method for update_articles.  However,
I think it works.

here's a brief illustration of the (admittedly slightly confusing) data
path:

after the article crawler crawls an article, it sends an update_article
message to the controller containing the contents of the crawled article;
this notes in the database that the article has been crawled and stores
its contents

then the article crawler shoots off a scrape_article message (including
the article contents) which is handled by the scraper service.  the scrape
is performed, and then the scraper sends another update_article message
to the controller, but this time with the scrape result, which is saved
in the database.  Controller.update_resource also at this point calls
the Controller._update_scraped_articles_hook--this extracts the site
information from the scrape results and saves a new site document if
necessary; if the site includes a new logo image it also shoots off
a crawl_image message which is handled by the image crawler.  This
handles downloading and saving the site's logo

eventually the controller, when receiving a update_article message
with scrape data would then also fire off an article_scraped event
to the event stream for recsystems; this could either include all
the scrape data outright, or just the ID of the scraped article,
which the recsystem can then fetch from the API.  I'm not sure I
really see any reason the article_scraped event couldn't include
all the scrape data though; in most cases it won't be that large?
Since recsystems will probably want to immediately access the
article data anyways, this would just save an API call.