[backend][refactoring] replace Flask with Quart for the HTTP API
Quart is a clone of Flask but with full asyncio support including support for
websockets

It should make it easier to integrate the API handlers with asyncio code, such
as sending/receiving messages via RabbitMQ

Unfortunately, the backend is still using Python 3.6, and the last version of
Quart to support Python 3.6 is a little behind Flask in terms of API (e.g. the
need to explicitly call jsonify() when sending JSON responses).  So this required
just some minor changes.  If we switch the backend to Python 3.7+ we can use a
newer version of Quart as well, which might be desireable at some point