[backend][refactoring] update the check_auth decorator so that it can work with alternative request globals
In Flask (as well as Quart) there is a thread-specific global call 'request'
which represents the current request being handled.  It's what we use to
check headers for authentication/authorization among other things.

Quart, however, provides a similar global called 'websocket'.  It has much
of the same interface as 'request', but is specifically for handling websocket
connections.  In order to provide auth on websocket endpoints, we need to
be able to tell check_auth when it should check auth headers on the 'websocket'
global instead of the 'request' global