[backend][task] start implementing the event_stream fanout exchange
this is the main "event stream" discussed in previous design documentation--the
stream of all events that are ultimately passed along to individual recsystems,
such as new articles, user events, user assignments, and the like

currently just two events are implemented: user->article interactions (clicks,
ratings, etc.) are sent on the event stream--currently this is a firehose, but
once user assignments for recsystems are implemented this may only send such
events to the recsystems assigned to a user

also NEW_ARTICLE events are produced every time an article is successfully
scraped--this are broadcast to all recsystems

recsystems connect to ws:///event_stream to connect to the event stream
websocket.  It's not clear to me yet that the EventStreamHandler class is
really necessary.  When I started writing it, it seemed like it might be
a useful place to collect certain logic, such as management of connected
recsystems.  Not 100% clear if it will truly be necessary though (might
just move that functionality onto the renewal_backend.web.App class).