[backend][refactoring] rename EventStreamHandler to RecsystemsHandler
this class is about more than just handling event stream events; it's really
more about managing connections from recsystems.  part of this involves
processing event stream events and converting them to RPC calls to the
targeted recsystems, but it may grow to do more than that as well

also adds a .rpc() method to it which will allow other request handlers to
directly make RPC calls to a recsystem (e.g. the recommendations() handler
will want to directly make an RPC call to a recsystem in order to retrieve
recommendations while handling the request