[mobileapp] change how different versions of app configs are loaded
previously this attempted to use the --release-channel command-line
option to determine which config file to use.  For example, when
running `expo build:android --release-channel=staging` the idea was
that this would use the config file `app.config.staging.json`.

It turns out this doesn't work--due to the details of how the config
file is read (it is actually done in a subprocess for some reason)
the command-line flags of the original command (e.g.
`expo build:android`) are not accessible to `app.config.js`.

However, any environment variables are.  So instead the config
variant is specified by an environment variable called `RENEWAL_ENV`.
So now we have to run something like:

    RENEWAL_ENV=staging expo build:android --release-channel staging

which is annoyingly redundant, but the only feasible workaround for
now.  Of course technically there's no reason the value of RENEWAL_ENV
and the name of the release channel have to be the same, but for
sanity's sake it would be a good idea to keep it that way.